Initial Commit of New Repo
I had some security issues I was worried about with my last Repo for
this project.  I had it private as at the time I did not plan to make
this a public repo.  I had a piece of sensitive data, and it seemed
easier to just make a new Repo and upload than to use the git shell to
purge the text.

So far this project has the core functionality I want in the site.  The
Schoolpage, send data to my Azure database via the review form and then
displays back reviews using the PHP code below it.  I am currently
working on the Register users and login functions of the site.