Changed Some Styling
I took a break from creating the user functionality such as registration
and login.  Instead I worked on the overall look of the site.  I added
the text in the menu for Colleges and Universities.  I added a gray
noise background, updated some fonts, and played around with drop
shadow.

Up next is user functionality.