Commit before I go back to Welland
Just a commit to save my minute work accomplished before I get going
back to Welland.

I've had some thoughts on the login/registration system.  Who is going
to want to create an account and verify it to post just one review?  I'm
thinking I will hold on to my login/registration stuff for later
development on the site.  Perhaps if the site is a success I can have
some other functionality.  I might just keep it in for fun.  Either way
I will keep going with session wide code in order to make it easy to add
in later.